Bug Fixes and AI Improvement
Added mechanic so that AI will only put player in check if either the checking will not be captured or it can be traded. In future, will add trade value mechanic to help AI decide whether or not to make a trade.

Fixed bug where straight ahead pawn moves were counted as contesting, or overtaking moves and AI would not move there